REFACTOR-#0000: remove code duplication for 'broadcast_apply' method
